+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support generic type for datatypes and functions.
 - Support file dump in output directory.
+- Implement lookahead branching heuristic.
 
 ## [0.1.3] - 2026-02-04